fix: upstream code, and fix expose_name removal
