+++ src/global/common.rs
@@ -1,10 +1,13 @@
-use super::{EnvironmentName, ExposedName};
+use super::{extract_executable_from_script, EnvironmentName, ExposedName, Mapping};
 use fancy_display::FancyDisplay;
 use fs_err as fs;
 use fs_err::tokio as tokio_fs;
+use indexmap::IndexSet;
+use itertools::Itertools;
 use miette::{Context, IntoDiagnostic};
 use pixi_config::home_path;
 use pixi_manifest::PrioritizedChannel;
+use pixi_utils::executable_from_path;
 use rattler_conda_types::{Channel, ChannelConfig, NamedChannelOrUrl, PackageRecord, PrefixRecord};
 use std::collections::HashMap;
 use std::ffi::OsStr;
@@ -405,6 +408,67 @@ pub(crate) fn channel_url_to_prioritized_channel(
         .into())
 }
 
+/// Figures out what the status is of the exposed binaries of the environment.
+///
+/// Returns a tuple of the exposed binaries to remove and the exposed binaries to add.
+pub(crate) async fn get_expose_scripts_sync_status(
+    bin_dir: &BinDir,
+    env_dir: &EnvDir,
+    exposed: &IndexSet<Mapping>,
+) -> miette::Result<(IndexSet<PathBuf>, IndexSet<ExposedName>)> {
+    // Get all paths to the binaries from the scripts in the bin directory.
+    let locally_exposed = bin_dir.files().await?;
+    let executable_paths = futures::future::join_all(locally_exposed.iter().map(|path| {
+        let path = path.clone();
+        async move {
+            extract_executable_from_script(&path)
+                .await
+                .ok()
+                .map(|exec| (path, exec))
+        }
+    }))
+    .await
+    .into_iter()
+    .flatten()
+    .collect_vec();
+
+    // Filter out all binaries that are related to the environment
+    let related_exposed = executable_paths
+        .into_iter()
+        .filter(|(_, exec)| exec.starts_with(env_dir.path()))
+        .map(|(path, _)| path)
+        .collect_vec();
+
+    // Get all related expose scripts not required by the environment manifest
+    let to_remove = related_exposed
+        .iter()
+        .filter(|path| {
+            !exposed
+                .iter()
+                .any(|mapping| executable_from_path(path) == mapping.exposed_name().to_string())
+        })
+        .cloned()
+        .collect::<IndexSet<PathBuf>>();
+
+    // Get all required exposed binaries that are not yet exposed
+    let to_add = exposed
+        .iter()
+        .filter_map(|mapping| {
+            if related_exposed
+                .iter()
+                .map(|path| executable_from_path(path))
+                .any(|exec| exec == mapping.exposed_name().to_string())
+            {
+                None
+            } else {
+                Some(mapping.exposed_name().clone())
+            }
+        })
+        .collect::<IndexSet<ExposedName>>();
+
+    Ok((to_remove, to_add))
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -505,5 +569,77 @@ mod tests {
         } else {
             assert_eq!(executable_script_path, path.join(exposed_name.to_string()));
         }
+    }
+
+    #[tokio::test]
+    async fn test_get_expose_scripts_sync_status() {
+        let tmp_home_dir = tempfile::tempdir().unwrap();
+        let tmp_home_dir_path = tmp_home_dir.path().to_path_buf();
+        let env_root = EnvRoot::new(tmp_home_dir_path.clone()).unwrap();
+        let env_name = EnvironmentName::from_str("test").unwrap();
+        let env_dir = EnvDir::from_env_root(env_root, &env_name).await.unwrap();
+        let bin_dir = BinDir::new(tmp_home_dir_path.clone()).unwrap();
+
+        // Test empty
+        let exposed = IndexSet::new();
+        let (to_remove, to_add) = get_expose_scripts_sync_status(&bin_dir, &env_dir, &exposed)
+            .await
+            .unwrap();
+        assert!(to_remove.is_empty());
+        assert!(to_add.is_empty());
+
+        // Test with exposed
+        let mut exposed = IndexSet::new();
+        exposed.insert(Mapping::new(
+            ExposedName::from_str("test").unwrap(),
+            "test".to_string(),
+        ));
+        let (to_remove, to_add) = get_expose_scripts_sync_status(&bin_dir, &env_dir, &exposed)
+            .await
+            .unwrap();
+        assert!(to_remove.is_empty());
+        assert_eq!(to_add.len(), 1);
+
+        // Add a script to the bin directory
+        let script_path = if cfg!(windows) {
+            bin_dir.path().join("test.bat")
+        } else {
+            bin_dir.path().join("test")
+        };
+
+        let script = if cfg!(windows) {
+            format!(
+                r#"
+            @"{}" %*
+            "#,
+                env_dir
+                    .path()
+                    .join("bin")
+                    .join("test.exe")
+                    .to_string_lossy()
+            )
+        } else {
+            format!(
+                r#"#!/bin/sh
+            "{}" "$@"
+            "#,
+                env_dir.path().join("bin").join("test").to_string_lossy()
+            )
+        };
+        tokio_fs::write(script_path, script).await.unwrap();
+
+        let (to_remove, to_add) = get_expose_scripts_sync_status(&bin_dir, &env_dir, &exposed)
+            .await
+            .unwrap();
+        assert!(to_remove.is_empty());
+        assert!(to_add.is_empty());
+
+        // Test to_remove
+        let (to_remove, to_add) =
+            get_expose_scripts_sync_status(&bin_dir, &env_dir, &IndexSet::new())
+                .await
+                .unwrap();
+        assert_eq!(to_remove.len(), 1);
+        assert!(to_add.is_empty());
     }
 }
